requirements.txt
Flask
flask_mysqldb
PyJWT
mysqlclient 
bcrypt 